+++ src/Enums/Gender.php
@@ -2,12 +2,12 @@
 
 namespace Saritasa\Enums;
 
-use Saritasa\Enum;
+use Saritasa\NamedEnum;
 
 /**
  * Human gender - Male or Female
  */
-class Gender extends Enum
+class Gender extends NamedEnum
 {
     const MALE = 'Male';
     const FEMALE = 'Female';