+++ src/Enums/PaymentType.php
@@ -2,12 +2,12 @@
 
 namespace Saritasa\Enums;
 
-use Saritasa\Enum;
+use Saritasa\NamedEnum;
 
 /**
  * How user pays money
  */
-class PaymentType extends Enum
+class PaymentType extends NamedEnum
 {
     const ANDROID_PAY = 'Android Pay';
     const APPLE_PAY = 'Apple Pay';