+++ src/Exceptions/InvalidEnumValueException.php
@@ -9,31 +9,42 @@ use Throwable;
  */
 class InvalidEnumValueException extends InvalidArgumentException
 {
-    /* @var array $possibleValues possible values */
-    protected $possibleValues;
+    /**
+     * The valid enum constants.
+     *
+     * @var array $validConstants
+     */
+    protected $validConstants;
 
     /**
      * Thrown, if the argument passed to the function does not match any of the possible values.
      *
-     * @param array $possibleValues possible values
-     * @param string|null $message exception message
-     * @param int $code http code
-     * @param Throwable|null $previous previous thrown exception
+     * @param string $invalidConstant The invalid constant name.
+     * @param array $validConstants The valid enum constants.
+     * @param string|null $message The exception message.
+     * @param int $code The exception code.
+     * @param Throwable|null $previous The previously thrown exception.
      */
-    public function __construct(array $possibleValues, string $message = null, int $code = 0, Throwable $previous = null)
+    public function __construct(
+        string $invalidConstant,
+        array $validConstants,
+        string $message = null,
+        int $code = 0,
+        Throwable $previous = null)
     {
-        $this->possibleValues = $possibleValues;
-        $message = $message ?? 'Value must be from the list: ' . implode(', ', $possibleValues);
+        $this->validConstants = $validConstants;
+        $message = $message ?? "Constant \"$invalidConstant\" does not exist. Valid values are " .
+            implode(', ', $validConstants);
         parent::__construct($message, $code, $previous);
     }
 
     /**
-     * Returns list of possible values.
+     * Returns the list of the valid enum constants.
      *
      * @return array
      */
-    public function getPossibleValues(): array
+    public function getValidConstants(): array
     {
-        return $this->possibleValues;
+        return $this->validConstants;
     }
 }